[refactor] alteração do campo valor de bigdecimal pra integer para atender ao teste automatizado

+++ src/main/java/br/com/ebanx/home/test/dto/BalanceDto.java
@@ -2,11 +2,9 @@ package br.com.ebanx.home.test.dto;
 
 import br.com.ebanx.home.test.entity.AccountEntity;
 
-import java.math.BigDecimal;
-
 public class BalanceDto {
     public Integer id;
-    public BigDecimal balance;
+    public Integer balance;
     public BalanceDto(AccountEntity accountEntity){
         this.id = accountEntity.getAccountCode();
         this.balance = accountEntity.getAmount();
